Stylistic changes for codacy
Made edits to reduce duplication detected by Codacy.
Some class methods were made into their own
functions where they made sense. In places where
a method was executing identical code we extracted
the code into its own method within the class.
Will also make the code more readable.
